feat(camera-info): CreateCameraInfo node + CameraInfoState widget input
