+++ nodes.py
@@ -2503,6 +2503,7 @@ async def init_builtin_extra_nodes():
         "nodes_moge.py",
         "nodes_mediapipe.py",
         "nodes_gaussian_splat.py",
+        "nodes_camera.py",
         "nodes_triposplat.py",
         "nodes_depth_anything_3.py",
         "nodes_seed.py",
